# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * News aggregator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NewsItem {
  id: number;
  title: string;
  summary: string;
  content: string;
  imageUrl: string;
  source: string;
  topic: string;
  region: string;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface NewsInput {
  title: string;
  summary: string;
  content: string;
  imageUrl: string;
  source: string;
  topic: string;
  region: string;
  /** @nullable */
  externalId?: string | null;
}

export interface ErrorResponse {
  message: string;
}

export interface MessageResponse {
  message: string;
}

export type ListNewsParams = {
/**
 * Comma-separated list of topics
 */
topics?: string;
/**
 * Region filter
 */
region?: string;
limit?: number;
offset?: number;
};