# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: News aggregator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: news
    description: News operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news items
      parameters:
        - name: topics
          in: query
          schema:
            type: string
          description: Comma-separated list of topics
        - name: region
          in: query
          schema:
            type: string
          description: Region filter
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of news items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /news/refresh:
    post:
      operationId: refreshNews
      tags: [news]
      summary: Trigger a background news refresh
      responses:
        "200":
          description: Refresh started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get a single news item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    NewsItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        source:
          type: string
        topic:
          type: string
        region:
          type: string
        externalId:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - title
        - summary
        - content
        - imageUrl
        - source
        - topic
        - region
    NewsInput:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        source:
          type: string
        topic:
          type: string
        region:
          type: string
        externalId:
          type: ["string", "null"]
      required:
        - title
        - summary
        - content
        - imageUrl
        - source
        - topic
        - region
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message