# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * News aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: number;
  title: string;
  summary: string;
  content: string;
  imageUrl: string;
  source: string;
  topic: string;
  region: string;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  createdAt?: string | null;
}